pyproject.toml
[project]
name = "hrr-language-memory-poc"
version = "0.1.0"
description = "Minimal HRR+AMM language memory proof of concept"
requires-python = ">=3.10"
dependencies = [
    "google-genai>=1.0.0",
    "numpy>=1.24",
    "pydantic>=2.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
